+++ src/agnocastlib/src/agnocast.cpp
@@ -3,8 +3,8 @@
 #include "agnocast/agnocast_ioctl.hpp"
 #include "agnocast/agnocast_mq.hpp"
 #include "agnocast/agnocast_version.hpp"
-#include "agnocast/bridge/agnocast_performance_bridge_manager.hpp"
-#include "agnocast/bridge/agnocast_standard_bridge_manager.hpp"
+#include "agnocast/bridge/performance/agnocast_performance_bridge_manager.hpp"
+#include "agnocast/bridge/standard/agnocast_standard_bridge_manager.hpp"
 
 #include <dlfcn.h>
 #include <sys/types.h>
