+++ src/agnocastlib/src/bridge/performance/agnocast_performance_bridge_ipc_event_loop.cpp
@@ -1,22 +1,21 @@
-#include "agnocast/bridge/agnocast_standard_bridge_ipc_event_loop.hpp"
+#include "agnocast/bridge/performance/agnocast_performance_bridge_ipc_event_loop.hpp"
 
 #include "agnocast/agnocast_mq.hpp"
 #include "agnocast/agnocast_utils.hpp"
 
-#include <unistd.h>
-
+#include <utility>
 #include <vector>
 
 namespace agnocast
 {
 
-StandardBridgeIpcEventLoop::StandardBridgeIpcEventLoop(const rclcpp::Logger & logger)
+PerformanceBridgeIpcEventLoop::PerformanceBridgeIpcEventLoop(const rclcpp::Logger & logger)
 : IpcEventLoopBase(
     logger,
     // 1. MQ Name
-    create_mq_name_for_bridge(getpid()),
+    create_mq_name_for_bridge(PERFORMANCE_BRIDGE_VIRTUAL_PID),
     // 2. Message Size
-    BRIDGE_MQ_MESSAGE_SIZE,
+    PERFORMANCE_BRIDGE_MQ_MESSAGE_SIZE,
     // 3. Block Signals
     {SIGTERM, SIGINT},
     // 4. Ignore Signals
