+++ src/agnocastlib/src/bridge/standard/agnocast_standard_bridge_loader.cpp
@@ -1,4 +1,4 @@
-#include "agnocast/bridge/agnocast_standard_bridge_loader.hpp"
+#include "agnocast/bridge/standard/agnocast_standard_bridge_loader.hpp"
 
 #include "agnocast/agnocast_utils.hpp"
 #include "agnocast/bridge/agnocast_bridge_utils.hpp"
