+++ src/agnocastlib/src/bridge/templates/a2r_bridge_plugin.cpp.em
@@ -2,7 +2,7 @@
 // This file was automatically generated by generate_a2r_bridge_plugin.py using EmPy.
 
 #include "agnocast/bridge/agnocast_bridge_node.hpp"
-#include "agnocast/bridge/agnocast_performance_bridge_plugin_api.hpp"
+#include "agnocast/bridge/performance/agnocast_performance_bridge_plugin_api.hpp"
 #include "agnocast/agnocast_subscription.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include <utility>
