+++ src/agnocastlib/src/bridge/templates/r2a_bridge_plugin.cpp.em
@@ -2,7 +2,7 @@
 // This file was automatically generated by generate_r2a_bridge_plugin.py using EmPy.
 
 #include "agnocast/bridge/agnocast_bridge_node.hpp"
-#include "agnocast/bridge/agnocast_performance_bridge_plugin_api.hpp"
+#include "agnocast/bridge/performance/agnocast_performance_bridge_plugin_api.hpp"
 #include "agnocast/agnocast_publisher.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include <utility>
